Preserve some TextEditor settings when language mode changes
This change fixes #13829 which reports that the `softWrapped` setting of
an untitled TextEditor is lost when the buffer is saved to a file.  This
is caused by logic that updates TextEditor settings when the buffer's
language mode changes.

The fix is to mark some editor settings as "preserved" so that they do
not get change when switching between two language modes that do not
have a different value for those preserved settings.